pyproject.toml
[project]
name = "fold"
version = "0.1.0"
description = "Fold your codebase into a single, easily digestible Markdown file for consumption by an LLM."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "pyyaml>=6.0.2",
    "tiktoken>=0.8.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.2",
]
